login validation

+++ src/main/resources/static/js/detail.js
@@ -1,24 +1,66 @@
 
 $(document).ready(function () {
 
-    $('#editBtn').on("click",function(){
-        let contents_section =  $('.edit');
-        let contents = $('#1-contents');
-        $('#editBtn').hide();
-        $('#deleteBtn').hide();
-        $('#sendBtn').show();
-        contents_section.show();
-        $('#1-textarea').val(contents.text());
-        $(contents).hide();
-    });
-
-
-
-
 })
+
+// 메시지 작성자랑 인가 체크
+function chkMessageUser(){
+    if(username == null){
+        alert("로그인이 필요합니다. 로그인페이지로 이동합니다.")
+        location.href="/user/login";
+        return false;
+    }else if(msgName !== username){
+        alert("권한이 없습니다.");
+        return false;
+    }else{
+        return true;
+    }
+
+}
+// comment 작성자랑 인가 체크
+function chkCommentUser(id){
+    let commentName = $('#'+id+'-commentName').text();
+    console.log(username);
+    console.log(commentName);
+
+    if(username == null){
+        alert("로그인이 필요합니다. 로그인페이지로 이동합니다.")
+        console.log("로그인 필요");
+        location.href="/user/login";
+        return false;
+    }else if(username !== commentName){
+        console.log("권한 없음");
+
+        alert("권한이 없습니다.");
+        return false;
+    }else{
+        return true;
+    }
+
+}
+// message section 인가 체크 및 버튼 스위칭
+function messageSwitchBtn(){
+    if(chkMessageUser()===false){
+        return false;
+    }
+    let contents_section =  $('.edit');
+    let contents = $('#1-contents');
+    $('#editBtn').hide();
+    $('#deleteBtn').hide();
+    $('#sendBtn').show();
+    contents_section.show();
+    $('#1-textarea').val(contents.text());
+    $(contents).hide();
+}
+// 메세지 아이디 변수 선언
 let stringMsgId = location.href.split("/")[5];
 let msgId = parseInt(stringMsgId);
+
+// 메시지 수정
 function updateMessage(){
+    if(chkMessageUser()===false){
+        return false;
+    }
     let msgContents = $('#1-textarea').val();
     let data ={'contents' : msgContents , 'name' : msgName, 'title':msgTitle};
     $.ajax({
@@ -37,7 +79,12 @@ function updateMessage(){
         }
     });
 }
+
 function deleteMessage(){
+
+    if(chkMessageUser()===false){
+        return false;
+    }
     $.ajax({
         type: "DELETE",
         url: `/api/contents/${msgId}`,
@@ -53,8 +100,14 @@ function deleteMessage(){
     });
 
 }
-
+// comment작성
 function createComt(){
+    if(username == null){
+        alert("로그인이 필요합니다. 로그인 페이지로 이동합니다.");
+        location.href='/user/login';
+        return false;
+    }
+
     let commentName = $('#commentName').text();
     let comment = $('#createComment').val();
     let data = {
@@ -81,7 +134,10 @@ function createComt(){
 
 // comment 버튼 스위칭
 function switchBtn(id){
-    let comment     =$('#'+id+"-comment");
+    if(chkCommentUser(id)===false){
+        return false;
+    }
+    let comment     = $('#'+id+"-comment");
     let commentArea = $('#'+id+"-commentArea");
     let editBtn     = $('#'+id+"-editBtn");
     let deleteBtn   = $('#'+id+"-deleteBtn");
@@ -92,6 +148,9 @@ function switchBtn(id){
 }
 // 코멘트 업데이트
 function updateComment(id){
+    if(chkCommentUser(id)===false){
+        return false;
+    }
     let commentEdit = $('#'+id+"-commentEdit");
     let data = {
             "comment" : commentEdit.val()
@@ -115,6 +174,9 @@ function updateComment(id){
 
 // 코멘트 삭제
 function deleteComment(id){
+    if(chkCommentUser(id)===false){
+        return false;
+    }
 
     $.ajax({
         type: "DELETE",
